requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
line-bot-sdk==3.7.0
requests==2.31.0
gspread==5.12.0
oauth2client==4.1.3
python-multipart==0.0.6
pydantic==2.4.2 